add all superficial variable and function mismatch fix to induced sampling
	This commit will also fix previous character encoding error
