--- tests/test.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python
-import matplotlib
-import nose
-matplotlib.use('agg')
-
-nose.main()
